+++ renderer/interface/IRenderer.hpp
@@ -16,7 +16,7 @@ namespace renderer {
         virtual bool Init() = 0;
         virtual void CreatePipeline(Material& mat, const char* vert_shader, const char* frag_shader, bool alpha = false) = 0;
         virtual void BeforeDraw() = 0;
-        virtual void Draw(RenderObject* first, int count) = 0;
+        virtual void Draw(RenderObject* first, size_t count) = 0;
         virtual void AfterDraw() = 0;
         virtual void WaitIdel() = 0;
         virtual void Release() = 0;
